chore(dataobj): move tenant a global construct
This follows up on the work started in #18843, moving the
section-specific `tenant` metadata field to storing tenancy info within
filemd.SectionInfo.

This allows the idea of sections being owned by a tenant to be a
dataobj-wide construct, and additionally permits deduplicating tenant
strings by making use of the existing string dictionary in
filemd.Metadata.

As of this commit, there is no enforcement that the tenant information
is provided for sections which are known to be single-tenant. That work
is left up to future PRs.