chore(dataobj)!: update indexpointers section to use internal/columnar
This commit updates the indexpointers section to extend the base
internal/columnar implementation.

As part of this change, the format version of the section has been
updated to be in sync with the columanr format version (2) and support
for old format versions has been removed.

indexpointersmd, which is now unused, has also been removed.