chore(dataobj)!: update logs section to use internal/columnar
This commit updates the logs section to extend the base
internal/columnar implementation.

As part of this change, the format version of the section has been
updated to be in sync with the columanr format version (2) and support
for old format versions has been removed.

logsmd, which is now unused, has also been removed.

The utility method to detect whether a column name was for metadata has
been replaced with a more general utility method for parsing column
types.

Since this change updated the SHA224 of the files, the temporary querier
tests needed to be updated to account for the different sharding
results.