refactor(dataobj): update all datasetmd usages to v2
This commit updates all datasetmd usages to use the v2 package, and
aligns the names with the names used in the new package.

This commit contains no changes to logic, though now that dataset.Values
are encoded using the v2 types, the sharding in dataobj/querier has once
again changed and the test assumptions needed to be updated to match the
changed SHA224.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>