chore(logql/bench): align data directory with production layout
Previously, the LogQL benchmarks put chunks and dataobjs in two
different directories. This prevented developers from being able to
properly run Loki locally against the synthetic data, as in a real Loki
setup, dataobjs are in a subdirectory of where chunks are stored.

This commit imitates that setup: put all chunk storage in in the root
data/ directory, then put dataobjs in data/dataobj.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>