chore(dataobj): track total download time of page metadata/data
We previously lacked a stat to be able to monitor the performance of
batched downloads of page metadata or data. Even if the number of
requested units to download remains the same, the mechanism through
which they are downloading can impact the total amount of time spent
waiting for the download.

This commit adds the stat as the total page download time, which
includes downloading the page data and the page headers.