chore(dataobj): decouple section composition from snapshot
dataobj.snapshot was built with the assumption that the metadata region
of a section always followed the data region of a section. This coupled
it to the encoder implemenetation, which encoded the locations inside
the metadata itself.

To make it easier to move to changing the layout (so that all metadata
regions are encoded together), dataobj.snapshot has been updated to more
generally have a "sectionRegion," which the snapshot has no knowledge of
beyond its scratch.Handle and size. On the encoder side, each section is
split into two sectionRegions: one for data, and one for metadata.

With this change, dataobj.snapshot encodes individual regions in the
order received. This allows encoder to update the ordering without
needing to update the snapshot code.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>