chore(dataobj/querier): remove fragile sharding tests
The sharding tests depended on the SHA-224 of the test objects, which
would change every time an update was made to the format. This resulted
in needing to change the test assertions.

Since dataobj/querier is only used in the LogQL benchmarks and will
never hit production (as it's for running data objects on the old
engine), there's little value in chasing the correctness of these tests,
so they have been removed for now.