chore(engine): fix Release/Retain semantics
Reference counting in Go can be hard to reason about. This commit
introduces consistent semantics with the following rules:

* Call Retain whenever there may be two active reference to the Record,
  such as when

  * a function accepts a Record as a parameter and stores it, or

  * a function returns a Record and continues to use it in future calls.

* Call Release when

  * the Record is dropped from local state and was previously Retained,
    or

  * when the Record is no longer needed and is not propagated to a
    caller.

Bsaed on these rules, most of our Pipelines do not need to call Retain
or Release, despite temporarily caching the Record for coordinating the
pair of Read/Value calls. As long as a Pipeline does not continue to use
the Record after it returns it to the caller, it does not need to call
Retain.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>