fix: deadlock when shutting down dataobj-consumers
This commit fixes a deadlock that would occur when shutting down
dataobj-consumers where the stopping() and OnPartitionsRevoked()
callback in service.go would both content for the same mutex.

We fix the deadlock by abstracting the management of processors
into a partition processor lifeycler, and the handling partition
assignment and revocation into a partition lifecycler. Meanwhile,
records are polled by a consumer, and then appended to the
correct partition processor.

The shutdown mechanism works by first shutting down the consumer
so no more records can be fetched. It then shutsdown the
lifecycler, which prevents new processors from being created
should there be a group rebalance while shutting down. Last it
shuts down the consumer and metatstore events clients. This
eliminates the need for a shared mutex.