chore: Add fixed ordering to edges in plan and only allow for single parent per plan node (#18978)
When creating a query plan, we now retain the order in which edges were added to the plan. Additionally, no node should ever have more than one parent, so the Plan struct and corresponding functions were simplified accordingly.

We want to order inputs in the query plan based on the time range of each input. While it's possible to group the inputs without any change to the Plan struct, there is no guarantee that the inputs will remain in order. To force the ordering to stay the same, we replace non-ordered nodeSet properties with []Node arrays.