chore(util): add utility for parallelizing ranged downloads
This introduces the rangeio utility package, which exposes a ReadRanges
function for efficiently reading multiple ranges in parallel.

Ranges which are close together (up to CoalesceSize) are coalesced into
a single, larger range. Coalesced ranges will not exceed MaxRangeSize.

The resulting coalesced ranges are read in parallel, up to a
configurable number of maximum workers. If the number of coalesced
ranges is fewer than this amount of parallelism, ranges are split to
increase throughput. The MinRangeSize determines the lower bound of how
small a range can be; ranges won't be split if the resulting halves
would be too small.

This logic is based on findings that througput to object storage is
maximized when:

* Using as much parallelism as one can tolerate without saturating the
  network.

* Parallelizing download ranges of 8MB.

* Parallelizing smaller download ranges if the number of ranges was
  smaller than the amount of workers.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>