feat(operator): deploy network policies with LokiStack **What this PR does / why we need it:** This PR introduces the ability to enable network policies. By enabling network policies the operator will deploy a set of policies with the goal of restricting the allowed communications to&by LokiStack components to only the ones strictly necessary
**Which issue(s) this PR fixes**:
https://issues.redhat.com/browse/LOG-7198

**Special notes for your reviewer**:
In this PR we have introduced the struck of `FeatureGate` which
goal is to tell the code if a certain feature should be enabled or not.
This was introduced as for the first time instead of adding
NetworkPolicies to the feature gate we are instead adding them to the
LokiStack API since they are something that should be enabled per
LokiStack deployment.