chore(logqlmodel/stats): print new engine summary in metrics.go
This change adds a new statistic for whether the V2 query engine was
used. When it is used, the printed list of key-value pairs in metrics.go
includes the stats for data objects.

Regardless of this change, these stats will not show up in Grafana's
query inspector, which has an allowlist[1] of stats to expose.

[1]:
https://github.com/grafana/grafana/blob/58e649a956b1b39364da5d371857ae28f5e4e0a4/pkg/tsdb/loki/frame.go#L314