refactor: move key collection into physical optimizer
This commit moves key collection needed for parsing out of the logical
planner and into a physical optimizer. This removes the need for the
expression resolver and column registry, which have been removed. This
also removes the need for any requested key tracking in the logical
parse node. As a further optimization, this PR also does some label
identification in the logical planning, as this makes key disambiguation
easier in the physical optimizer.

Finally, thie PR aslo removes all logic releated to unwrapping values,
which includes casting and type hints. That will be down in a later PR.