feat: add more specificity to affinity and topology spread constraints (#17404)
This commit allows templating in the affinity and topology spread constraints, and updates the values to be more specific about the app.kubernetes.io/instance and app.kubernetes.io/name kubernetes recommended label.

When deploying into a cluster with a low number of nodes, or a cluster with the app.kubernetes.io/component label already used extensively, you cannot place these pods. If we want them to be scheduled apart, we should be more specific with the label selector.

This follows on from the draft PR https://github.com/grafana/loki/pull/17405