chore(engine): improve performance of matching window calculation (#19013)
Currently range aggregation linearly scans through all the possible time windows (a 24h query with a 1m step results in 1440 windows) to find the matching ones. This is quite expensive as this matching is done for each row.

This PR updates the matching code to use arithmetic which runs in constant time for cases where step >= range and for overlapping windows (step < range) it uses binary search.

Co-authored-by: Ashwanth Goli <iamashwanth@gmail.com>
Signed-off-by: Christian Haudum <christian.haudum@gmail.com>