feat: add replicas to loki-canary deployment
Currently, there's no middle ground between deploying the canary as a single
replica or deploying to your entire cluster. For very large clusters, this
can result in a lot of pressure on Loki. But using a single canary replica
doesn't give a huge amount of confidence on the health of the deployment.

Default to 1 replica so this is a no-op change for existing setups.