chore(engine): Simplify `Pipeline` interface (#19110)
This PR replaces the two calls to retrieve a value from the pipeline with a single one.
To do this, the `Read()` method now returns two arguments: `arrow.Record` and `error`, rather than just the error.
If the error is not `nil`, the first argument (arrow.Record) must be ignored by the caller.
The `Value()` method is not needed any more and was removed.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>