chore(engine): remove cache size limit when starting a scan
Only allowing a subset of pages to be prefetched helped reduce peak
memory usage, but came at the cost of scan latency being way too high.

This commit removes the prefetch limit for a section scan, allowing all
unpruned pages to be requested when first uncached page is requested.
This does not change the behaviour of clearing the cache; advancing the
reader past a cached page will continue to remove it from memory.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>