chore(dataobj): make prefetch opt-in
Prefetching pages is unnecessary when the dataset is already held in
memory. This is the case for the logs section builder, where the
intermediate (in-memory) tables are read for merging.

All other readers currently force prefetching. If other readers start to
be used with in-memory data, we can further propagate the configuration
option.

This change will help slightly reduce the overhead of object building.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>