chore(engine/logical): fix order/duplicates of SSA form
This fixes the SSA form of logical plans to follow linearly, such that
no instruction or value references a value (by ID) declared after it.
This was caused by predicates on MAKETABLE not being added as
instructions when they were first encountered.

Fixing this also revealed that an instruction/value with two references
was generating a new ID for each time it was walked. This is also fixed
by only generating an ID if one does not already exist.

Fixing these issues does not change the behaviour of the logical plan;
it just makes the SSA printed form more consistent with how SSA forms
are typically formatted.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>