chore(dataobj): Fix panic in column builder (#19153)
There is an edge case in the column builder where it panics when the AppendNulls() call tries to append more NULL values than the specified MaxPageRows setting when backfilling rows.
In that case the builder would panic because canAppend would always return false even when the previous page was flushed.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>