fix: fix context canceled stops reader from cleaning up
This commit fixes a bug where cancelling the context when there
are no more records prevented the reader from cleaning up.
For example, when changing from starting to running phase
after processing consumer lag at startup. This is a problem
for dataobj-consumers which might want to build an object
and flush a metastore event.