chore: use registerers with prefix for kafka metrics
This commit updates all loki_kafka_client prefixed metrics to
derive their prefix from a wrapped registerer instead of using
the global prefix loki_kafka_client.

This is important as we will have two services declaring these
metrics at the same time: ingesters and dataobj consumers; and
we want each service to use their own metrics prefix.

The commit is expected to be a no-op and should not rename any
existing metrics. Instead, it just updates the code to derive
the prefix from the registerer instead.