chore: inject consumer group into offset manager
This commit contains another intermediate change as part of a
wider effort to make the offset manager independent of the
Kafka configuration. Instead of passing a kafka.Config to
figure out the consumer group we pass it instead. This leaves
just the topic to extract which we will do in another commit.