chore(engine): Add ColumnCompat node to physical plan
The `ColumnCompat` node is used to guarantee compatibility with the old
engine.
In the new engine it is possible to have the same column name but from
different sources, such as labels or structured metadata.

While the old engine suffixes colliding names with a `_extracted`, the
new engine returns them in both the labels and structured metadata
response.

This new node is used to keep the old behaviour without implementing the
logic directly into the engine, but having it separate, so it can
easliy be disabled again.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>