chore(engine): Add `LOGQL_COMPAT` marker to logical plan (#19488)
The marker operation `LOGQL_COMPAT` indicates that the query is expected to return results that are compatible with the old v1 chunks engine.

It is used in the physical planner to add required `ColumnCompat` nodes to the plan that handle columns with colliding names but of different semantic types (label, metadata, parsed).