chore(engine): add Parallelize hint node
The new Parallelize node represents a hint to the engine to partition
and parallelize the children branches of the Parallelize and emit
results as a single sequence with no guaranteed order.

Parallelize is intended to be used by the scheduler for generating
tasks, but will not end up in the individual task plans.

If a physical plan containing the Parallelize hint node is executed, the
Parallelize node is ignored, and execution passes directly to its sole
input.

This commit does not contain any optimization passes that make use of
the Parallelize node. Future optimization passes can be written to "push
down" operations inside parallelization for work distribution.