chore(engine): add ScanSet node
ScanSet is a single node which represents a set of targets to scan in a
single operation.

As part of this change, a TopK can no longer be injected at the scan
level. Instead, processing a SORT from the logical plan directly results
in a TopK. buildNodeGroup and overlappingShardDescriptors are now unused
and have been removed.

The scheduler can use ScanSet to manually create smaller per-section
tasks.