fix: LineFilterLabelFilter.String() regexp correct delimiters (#17129)
When the querysharder executes, it converts a query into an AST representation before forwarding it on to the queriers. Unfortunately, when this AST representation is converted back into a string, it does not handle the case where a regexp can contain backticks - it always tries to wrap the regexp in backticks, which can lead to parser errors. This PR updates the String() method that converted the filter back into a query string and adapts the delimiters based on the content of the regexp.

(cherry picked from commit 0303aeebf18ab5ede4aa1ebb628576a669d9e38a)