chore(engine): introduce scheduler
The new scheduler package provides an implementation of workflow.Runner
that can assign tasks to a set of connected workers. Workers connect and
communicate through the interfaces used in the wire package.

See the wire package for details on the messages that will be sent
between the scheduler and workers.

This does not yet provide an implementation of a worker, which will be
done later.