fix(engine): fix deadlock between scheduler and workflow
The scheduler and the workflow can recursively invoke each others
methods, which can lead to a deadlock if one invokes the other while a
lock is held.

This adds a "notifier" to the scheduler to collect invocations of
workflow methods to be run on function exit after locks are released.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>