chore(engine): only cancel non-terminated children tasks
Previously, the workflow would respond to a terminated task by
requesting the cancellation of all of its children tasks. However, those
children tasks may have (and usually are) already terminated. When this
happens, the scheduler no longer recognizes those tasks.

This resulted in a lot of warn messages from the workflow.

Changing the workflow to only request the cancellation of non-terminated
tasks allieves the issue and removes the flood of warn messages.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>