perf(engine): Improve regexp expression evaluation (#19644)
This PR changes the evaluation of regex-match binary expression evaluations so that the regexp on right hand side of the expression is only compiled once for each batch evaluation, rather than for each row.

This saves a ton of allocations and improves general performance of queries with a regexp line or label filter.