chore(engine): update physical nodes to use ULID as ID
Updates all physical nodes to use a ULID as their ID, and makes the
field public for explicit node construction (which will be used for
protobuf conversion).

Unit tests which previously explicitly set the ULID have been updated to
leave the ID as the empty ULID.

Currently this field is never set (but will be in the following commit).