chore(engine): serializable physical plans (#19672)
This adds two new packages, `expressionpb` and `physicalpb`, which are
serializable representations of `physical.Expression` and `physical.Plan`,
respectively.

These packages include utility functions to convert between the protobuf
representations and the planner types.

A translation layer is used due to the complexity of integrating protobuf
throughout the engine, as well as difficulties with finding a clean pattern to
construct node types. 

Once we're able to identify a clean and consistent pattern for interacting with
the protobuf types throughout the codebase, the translation layer can be
removed in favour of using the protobuf types directly.

Co-authored-by: Sophie Waldman <sophie.waldman@grafana.com>