chore(engine): determine proper advertise address for scheduler
The address of the HTTP server listener cannot be used as the advertise
address, since listening on typical addresses like `0.0.0.0` don't
resolve into an address that other peers can use for connecting.

Instead, this chaneg adds interface names as a field, much like what's
used in the query-frontend configuration for the same purpose. The
interface names are used to infer a public address for advertising.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>