chore(engine): remove reference to internal type in exported function
Idiomatically, exported identifiers should not publicly reference
unexported identifiers. In this case, the public Handler method accepted
a non-exported type as an argument. This makes public APIs harder to
consume (callers must navigaate to the source code to understand what
implemenets that type).

In this case, we don't want callers providing anything other than the
concrete type, so we introduce a new non-exported function that accepts
the non-exported interface for testing.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>