chore(engine): make stream/task debug logging configurable
Logging every state change for tasks and streams is incredibly verbose,
but still potentially veluable when debugging issues with the engine.

Instead of taking out these log lines altogether, the verbose logging is
configurable as a limit per-tenant.

The new options are named generically to leave room for additioanl task-
or stream-level debugging functionality within the new engine.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>