chore(engine): Fix conversion from logical to physical Literal in math expressions (#19853)
Since both logical.Literal and physical.LiteralExpr embed the types.Literal interface, they both automatically also implement the interface.

This is problematic when creating a physical LiteralExpr from a logical Literal, rather than from the underlying value.

The serialization of the physical plan fails, because the pointer value of the interface is then a logical.Literal, rather than a physical.LiteralExpr