refactor(engine): update workflow to register all tasks/streams at once
This PR updates workflow.Runner to a manifest model:

* Workflows define the set of tasks and streams they may use upfront as
  a "manifest."
* Workflows manually unregister the manifest when the workflow itself is
  closed.

This change simplifies some state management and removes some edge cases
from the code; for example, it is no longer possible to create a task
for which there is already a listener (from a call to Listen).

This also makes it possible to introduce metrics in the scheduler to
report tasks by all state combinations (for as long as the associated
workflow is live). Previously, tasks would automatically be unregistered
from the scheduler when they reached a terminal state.

As part of this change, the call to Listen has been updated to provide
an interface to write to. This change pairs better with the manifest
model, as handlers for state changes must be provided alongside the
manifest.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>