chore(engine): define WorkerHello message
Defines a new message, WorkerHello, which is used to establish a
connection for control plane purposes, along with the total number of
threads that worker has available to use. This commit does not yet use
this new message type.

WorkerHello is not intended to mark worker threads as ready. It is
informational but can also be used for validation (such as rejecting a
WorkerReady message if the worker is exceeding its total number of
threads).