chore(engine)!: track worker connection type and thread count
Update the scheduler to track the worker connection type as either:

* Initial (the worker has just connected)
* Control plane (the connection is for communicating task status)
* Data plane (the connection is for communicating stream data)

The connection type is set to "control plane" when the worker sends a
WorkerHello message. This also sets the maximum number of threads for
the connection (currently unused).

The connection type is set to "data plane" when the worker sends a
StreamData message.

Connections may only transition from the initial type to a non-initial
type. Messages that are not expected based on the connection's state are
rejected.

Initially, the connection type helps ensure that we keep control plane
and data plane messages separately. It will also eventually be used for
observability information on the scheduler to track ongoing connections
and their purpose.

BREAKING CHANGE: After this change, both schedulers and workers must be
running the same code to properly communicate.