chore: implement UpdateRates with simple in-memory tracker
This commit implements UpdateRates with a simple in-memory tracker.
It does not replicate to Kafka currently because this produces
a huge amount of volume that we need to replay on every restart.
For now, since rate buckets are refilled very quickly, I propose
we keep these in-memory without replication while we think about
how to solve this.