chore(engine): send structured errors over the wire
This updates the wiring of the scheduler to communicate structured error
messages, including an HTTP status code along with an error message.
Schedulers and workers can use this to communicate when an error is
actionable, such as retrying a message later.

The high-level API of wire.Peer still uses `error`, but these can be
converted on-demand into a wire.Error for deeper analysis.

This change reserves the previous field, which was only used for error
messages without any detail. Once this is applied, scheduler peers will
not understand error messages from peers running the older code (though
they will still recognize that an error has occurred).

Signed-off-by: Robert Fratto <robertfratto@gmail.com>