chore(engine): have workers return HTTP 429 when no threads are available
Previously, workers would panic if they received a task assignment when
no threads were available, as they assumed a scheduler would only send a
task after one task was available.

As we move to a model where schedulers internally track how many threads
they can use, a worker may run out of threads if multiple schedulers are
assigning tasks to the same worker in parallel. To prepare for this
change, workers will now instead return an HTTP 429 back to the
scheduler when they are out of threads.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>