chore(engine): update how scheduler tracks readiness
Previously, the scheduler would track worker threads by placing the
worker connection into a FIFO queue for each WorkerReady message
received.

However, this forces workers to assign specific threads to schedulers,
which may dramatically reduce total compute capacity when only a subset
of schedulers are busy and as the number of schedulers increases. A
change to allow schedulers to share all worker threads is needed to
avoid this problem.

Now, ready workers are placed into a set, and a ready worker is selected
randomly. Ready workers are only removed from the set once the worker
responds with an HTTP 429 on task assignment, and placed back on the set
after the next WorkerReady.

This changes the purpose of WorkerReady from "exactly one thread is
ready" to "at least one thread is ready."

Signed-off-by: Robert Fratto <robertfratto@gmail.com>