chore(dataobj): stop preallocating memory for page builders
Initially, newPageBuilder would preallocate a buffer for the values
based on the maximum page size. This was done in an attempt to prevent
bytes.Buffer from growing slices beyond the configured maximum page
size.

However, this added a significant amount of memory overhead for datasets
with a large number of sparsely filled columns.

As it's more common for columns to be partially filled (especially for
structured metadata), it's more efficient to remove the preallocation
and let Go grow the buffers naturally.

This comes at the cost of full pages using slightly more memory than the
page size, usually up to the next power of two, if the target page size
doesn't align with a power of two.