fix: Handle endpoint without schema
With the v1 AWS SDK it was possible to specify an endpoint without
schema, which then chose the correct schema after setting
config.WithDisableSSL().

However, with the new SDK, setting `EndpointOptions.DisableHTTPS` to
true/false is not enough and yields the error:

```
 resolve endpoint: endpoint rule error, Custom endpoint `s3.eu-central-1.amazonaws.com` was not a valid URI
 ```

This commit adds the schema http or https to the endpoint depending on
the value of the field `insecure`.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>