fix: Fix regression in S3 client configuration (#20110)
### Summary

This PR fixes a regression in how Loki's S3 configuration options are handled to create the S3 client which was introduced with the upgrade of the AWS SDK from v1 to v2 (#19205).

With the v1 SDK the `s3` field of the `aws` configuration block was parsed to extract credentials, region, host, and schema.

```
storage_config:
  aws:
    s3: s3://<accesskey>:<secret>@<endpoint_or_region>/<bucketnames>
```

After the upgrade, only credentials were extraced, leaving all other parts of the URL as default, which would require explicit overrides.

This was a breaking change and not documented in the upgrade notes. We want to keep backwards compatibility, even though the Thanos object client is going to supersed the custom S3/GCS/... clients.

ref: #19908

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>