feat: add downscalePermittedFunc to check downscale is permitted
This commit adds a new abstraction downscalePermittedFunc which
returns true if a downscale is permitted. It adds a function
newChainedDownscalePermittedFunc that allows multiple funcs to
be chained, and a newOffsetCommittedDownscaleFunc that checks
if all records have been committed up to the end offset.

This prevents consumers from shutting down before processing all
records in a partition as this would lead to data loss.