chore(engine): add new WorkerSubscribe message
This adds a new message, WorkerSubscribe, used by the scheduler to
explicitly request when a worker has at least one worker thread
available.

Workers that receive a WorkerSubscribe message *must* send a WorkerReady
to the subscribed scheduler the next time a worker thread is available,
or if a worker thread is already available. Sending a WorkerReady
message in response to a WorkerSubscribe should clear the subscription
to reduce message noise.

IThis commit only updates the scheduler to send the message, but workers
are not yet updated to act on it.

ntroducing this message is backwards-compatible, as schedulers can
still receive WorkerReady messages without a subscription.