chore(engine)!: Share worker threads across all scheduler connections
Previously, worker threads would be assigned to a random connected
scheduler, with that assignment sticking until:

* That thread completes a task from the assigned scheduler, or
* the scheduler disconnects.

This is fine if there's only one scheduler, but it causes issues when
there are multiple schedulers:

* Idle schedulers may "hog" a thread that could be used by a busy
  scheduler.

* While the random selection is even over time, any given instant in
  time does not have even distribution of threads. In the worse case
  some schedulers may have zero threads assigned to them.

This is a particular problem as the total compute capcity per scheduler
decreases as the number of schedulers go up, even if queries are
distributed amongst the schedulers.

This commit fixes this issue by allowing any scheduler to give a task to
any ready worker thread. This permits scaling schedulers independently
of the number of workers without risking saturation.

Details
-------

A new message type, WorkerSubscribe, is introduced. This message is sent
by a scheduler to a worker, asking the worker to send a WorkerReady when
there is at least one ready thread.

WorkerSubscribe is sent by the scheduler when a worker connects, and
after a worker sends an HTTP 429 after rejecting a task assignment.

A new mechanism, `jobManager`, is used by the worker to bridge the
connection to schedulers and running worker threads. It is implemented
as a cancellable condition variable.

BREAKING CHANGE: Workers now expect schedulers to send WorkerSubscribe
before any WorkerReady message is sent.

Signed-off-by: Robert Fratto <robertfratto@gmail.com>